complete various todos

+++ coordinator/tributary/src/mempool.rs
@@ -114,7 +114,7 @@ impl<D: Db, T: TransactionTrait> Mempool<D, T> {
     tx: Transaction<T>,
     schema: &N::SignatureScheme,
     unsigned_in_chain: impl Fn([u8; 32]) -> bool,
-    commit: impl Fn(u32) -> Option<Commit<N::SignatureScheme>>,
+    commit: impl Fn(u64) -> Option<Commit<N::SignatureScheme>>,
   ) -> Result<bool, TransactionError> {
     match &tx {
       Transaction::Tendermint(tendermint_tx) => {
